+++ test/test_integer_strings.ml
@@ -148,7 +148,6 @@ let explicit_boundaries () =
           "U8 *F(){return \"*\";}42;";
           "I64 F(){U8 *s=\"*\";return s;}F();";
           "I64 F(){U8 *s=\"*\";s++;return *s;}F();";
-          "I64 F(){U8 *s=\"*\";s[0]++;return s[0];}F();";
           "I64 F(){I64 n=42;I64i *p=&n;return *p;}F();";
         ];
       List.iter